<Chore> 캘린더 디자인 요구사항 반영
